// packages/server/src/plugins-trigger/types.ts
export interface GitHubTriggerConfig {
  plugin: 'github';
  events: Array<'push' | 'pull_request'>;
  repository?: string; // glob e.g. 'acme/*' or 'acme/infra'
  branch?: string; // glob e.g. 'main', 'release/*'
  actions?: Array<'opened' | 'synchronize' | 'closed' | 'reopened'>; // PR only
}

export interface PerforceTriggerConfig {
  plugin: 'perforce';
  depot?: string; // glob e.g. '//depot/infra/*'
  user?: string; // glob
  client?: string; // glob
}

export type TriggerConfig = GitHubTriggerConfig | PerforceTriggerConfig;